Major changes

+++ PhotoGalleryAPI.Back/PhotoGalleryAPI/Controllers/AlbumsController.cs
@@ -1,11 +1,13 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using PhotoGalleryAPI.BaseResponse;
 using PhotoGalleryAPI.BaseResponse.Responses;
 using PhotoGalleryAPI.DAL.Entities;
 using PhotoGalleryAPI.Services.Services;
 using PhotoGalleryAPI.Shared.DTOs;
+using PhotoGalleryAPI.Storage;
 using System.Security.Claims;
 
 namespace PhotoGalleryAPI.Controllers
@@ -15,17 +17,19 @@ namespace PhotoGalleryAPI.Controllers
     public class AlbumsController : ControllerBase
     {
         private readonly IAlbumService _albumService;
+        private readonly IPhotoService _photoService;
 
-        public AlbumsController(IAlbumService albumService)
+        public AlbumsController(IAlbumService albumService, IPhotoService photoService)
         {
             _albumService = albumService;
+            _photoService = photoService;
         }
 
         [HttpGet]
         public async Task<ActionResult<IResponse<IEnumerable<AlbumDTO>>>> GetAll([FromQuery] int itemsPerPage = 1, [FromQuery] int selectedPage = 1)
         {
             var response = await _albumService.GetAllAsync(null, itemsPerPage, selectedPage);
-            return response.Success ? Ok(response) : BadRequest(response);
+            return Ok(response);
         }
 
         [HttpGet("My")]
@@ -36,7 +40,7 @@ namespace PhotoGalleryAPI.Controllers
                 throw new ArgumentException("Invalid person ID argument") :
                 await _albumService.GetAllAsync(x => x.CreatedByPersonId == Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)), itemsPerPage, selectedPage);
 
-            return response.Success ? Ok(response) : BadRequest(response);
+            return Ok(response);
         }
 
         [HttpPost]
@@ -44,15 +48,30 @@ namespace PhotoGalleryAPI.Controllers
         public async Task<ActionResult<IResponse<AlbumDTO>>> Create([FromBody] AlbumDTO albumDTO)
         {
             var response = await _albumService.AddAsync(albumDTO);
-            return response.Success ? Ok(response) : BadRequest(response);
+            return Ok(response);
         }
 
         [HttpDelete]
         [Authorize]
         public async Task<ActionResult<IResponse<AlbumDTO>>> Delete([FromQuery] string albumId = "")
         {
+            var photoResponse = await _photoService.GetAllAsync(x => x.AlbumId == Guid.Parse(albumId));
+
+            if(!photoResponse.Success)
+                return Ok(photoResponse);
+
+            foreach (var photo in photoResponse.Data)
+                await PhotoStorage.DeleteFileAsync(photo.Filename);
+
             var response = await _albumService.DeleteAsync(albumId);
-            return response.Success ? Ok(response) : BadRequest(response);
+            return Ok(response);
+        }
+
+        [HttpGet("Album/{albumId}")]
+        public async Task<ActionResult<IResponse<AlbumDTO>>> GetById(string albumId = "")
+        {
+            var response = await _albumService.GetByIdAsync(albumId);
+            return Ok(response);
         }
     }
 }
